requirements.txt
numpy 
pandas 
scikit-learn 
joblib 
matplotlib 
seaborn
category_encoder 
xgboost
fastapi 
pydantic
streamlit 
requests
uvicorn